requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0

# OpenAI - updated version to fix proxies error
openai>=1.54.0
httpx>=0.27.0

# LangSmith for tracing
langsmith>=0.1.0

# gRPC for ASR (optional)
grpcio>=1.62.0
grpcio-tools>=1.62.0
protobuf>=4.25.0

# Utilities
aiofiles==23.2.1
pydantic>=2.0.0,<3.0.0
typing-extensions>=4.9.0

redis==5.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
